+++ frontend/src/utils/copy.ts
